+++ src/spec/SceneSpec.ts
@@ -40,8 +40,8 @@ describe('A scene', () => {
       expect(scene.children.length).toBe(1);
    });
 
-   it('cannot have the same Timer added to it more than once', () => {
-      //TODO
+   xit('cannot have the same Timer added to it more than once', () => {
+      // TODO
    });
 
    it('cannot have the same TileMap added to it more than once', () => {
@@ -158,6 +158,91 @@ describe('A scene', () => {
       scene._initialize(engine);
 
       expect(initializeCount).toBe(1, 'Scenes can only be initialized once');
+   });
+
+   it('will update Actors that were added in a Timer callback', () => {
+      var updated = false;
+      var initialized = false;
+      var actor = new ex.Actor();
+      actor.on('initialize', () => {
+         initialized = true;
+      });
+      actor.on('postupdate', () => {
+         updated = true;
+
+         expect(initialized).toBe(true, 'Actor was not initialized before calling update');
+      });
+      actor.on('postdraw', () => {
+         expect(updated).toBe(true, 'Actor was not updated before calling draw');
+         expect(initialized).toBe(true, 'Actor was not initialized before calling draw');
+      });
+
+      // create Timer
+      var timer = new ex.Timer(() => {
+         scene.add(actor);
+      }, 10, false);
+
+      scene.add(timer);
+      scene.update(engine, 11);
+      scene.draw(engine.ctx, 11);
+
+      expect(scene.children.indexOf(actor)).toBeGreaterThan(-1, 'Actor was not added to scene');
+      expect(initialized).toBe(true, 'Actor was not initialized after timer callback');
+      expect(updated).toBe(true, 'Actor was not updated after timer callback');
+   });
+
+   it('will update UIActors that were added in a Timer callback', () => {
+      var updated = false;
+      var initialized = false;
+      var actor = new ex.UIActor();
+      actor.on('initialize', () => {
+         initialized = true;
+      });
+      actor.on('postupdate', () => {
+         updated = true;
+
+         expect(initialized).toBe(true, 'UIActor was not initialized before calling update');
+      });
+      actor.on('postdraw', () => {
+         expect(updated).toBe(true, 'UIActor was not updated before calling draw');
+         expect(initialized).toBe(true, 'UIActor was not initialized before calling draw');
+      });
+
+      // create Timer
+      var timer = new ex.Timer(() => {
+         scene.add(actor);
+      }, 10, false);
+
+      scene.add(timer);
+      scene.update(engine, 11);
+      scene.draw(engine.ctx, 11);
+
+      expect(scene.uiActors.indexOf(actor)).toBeGreaterThan(-1, 'UIActor was not added to scene');
+      expect(initialized).toBe(true, 'UIActor was not initialized after timer callback');
+      expect(updated).toBe(true, 'UIActor was not updated after timer callback');
+   });
+
+   it('will update TileMaps that were added in a Timer callback', () => {
+      var updated = false;
+      var tilemap = new ex.TileMap(0, 0, 1, 1, 1, 1);
+      tilemap.on('postupdate', () => {
+         updated = true;
+      });
+      tilemap.on('postdraw', () => {
+         expect(updated).toBe(true, 'TileMap was not updated before calling draw');
+      });
+
+      // create Timer
+      var timer = new ex.Timer(() => {
+         scene.add(tilemap);
+      }, 10, false);
+
+      scene.add(timer);
+      scene.update(engine, 11);
+      scene.draw(engine.ctx, 11);
+
+      expect(scene.tileMaps.indexOf(tilemap)).toBeGreaterThan(-1, 'TileMap was not added to scene');
+      expect(updated).toBe(true, 'TileMap was not updated after timer callback');
    });
 
 });